@@ -1,9 +1,9 @@
-from typing import Any, Literal, Optional
+from typing import Literal
 
 from sqlalchemy import JSON, Enum, String
-from sqlalchemy.orm import DeclarativeBase, Mapped, column_property, declared_attr, mapped_column
+from sqlalchemy.orm import Mapped, mapped_column
 
-from merchants.providers import factory
+from merchants.integrations import local_factory
 
 PaymentStatus = Literal["Created", "Paid", "Rejected"]
 
@@ -18,18 +18,19 @@ class PaymentMixin:
         nullable=True,
         default=None,
     )
-    provider_response: Mapped[JSON] = mapped_column(JSON, nullable=True, default=None)
-    provider_error_response: Mapped[str] = mapped_column(String(255), nullable=True, default=None)
+    provider_response: Mapped[JSON | None]
+    provider_error_response: Mapped[str | None]
     payment_status: Mapped[PaymentStatus] = mapped_column(
         Enum("Created", "Paid", "Rejected", name="payment_status_choices"),
         nullable=False,
         default="Created",
+        index=True,
     )
-    payment_payload: Mapped[JSON] = mapped_column(JSON, nullable=True, default=None)
+    payment_payload: Mapped[JSON | None]
 
-    name_column: Optional[str] = "name"
-    email_column: Optional[str] = "email"
+    name_column: str | None = "name"
+    email_column: str | None = "email"
 
     def make_payment(self):
-        provider = factory(self.provider)
+        provider = local_factory(self.provider)
         return f"{provider}"
