@@ -1,10 +1,11 @@
 from typing import Literal, get_args, get_origin
 
-from merchants.core.orm import sqla
+from merchants.orm import sqla
 
 
 def test_Payment_status_type():
     assert get_origin(sqla.PaymentStatus) is Literal
+
 
 def test_Payment_status_values():
     expected_values = ["Created", "Paid", "Rejected"]
